split alert component into alert and inlineAlert components. Adjust examples
