+++ src/core/Alert/index.ts
@@ -1,0 +1,2 @@
+export { Alert, AlertProps } from './Alert/Alert';
+export { InlineAlert, InlineAlertProps } from './InlineAlert/InlineAlert';
